moved problem-set-filter, added docstring
